fix: 2 回目以降にエラーが発生しないようにする

+++ set_docker_bash_completion.sh
@@ -10,6 +10,6 @@
 
 # https://docs.docker.com/docker-for-mac/#install-shell-completion
 etc=/Applications/Docker.app/Contents/Resources/etc
-ln -s $etc/docker.bash-completion $(brew --prefix)/etc/bash_completion.d/docker
-ln -s $etc/docker-machine.bash-completion $(brew --prefix)/etc/bash_completion.d/docker-machine
-ln -s $etc/docker-compose.bash-completion $(brew --prefix)/etc/bash_completion.d/docker-compose
+[[ -e $(brew --prefix)/etc/bash_completion.d/docker ]] || ln -s $etc/docker.bash-completion $(brew --prefix)/etc/bash_completion.d/docker
+[[ -e $(brew --prefix)/etc/bash_completion.d/docker-machine ]] || ln -s $etc/docker-machine.bash-completion $(brew --prefix)/etc/bash_completion.d/docker-machine
+[[ -e $(brew --prefix)/etc/bash_completion.d/docker-compose ]] || ln -s $etc/docker-compose.bash-completion $(brew --prefix)/etc/bash_completion.d/docker-compose
